refactor(playground): extract shell/ (rAF loop + resetAll)
Move all remaining orchestration code from main.ts into shell/:

shell/boot.ts — async init, wasm load, permalink decode
shell/wire-ui.ts — UiHandles, DOM handle collection
shell/apply-permalink.ts — applyPermalinkToUi, speed/intensity labels
shell/reset.ts — resetAll, configureTraffic, drainSeedBatch
shell/loop.ts — rAF render loop, updateScoreboard
shell/listeners.ts — attachListeners, attachKeyboardShortcuts
shell/state.ts — State interface

main.ts is now 3 lines: import + boot(). The 2002-line monolith
is fully dismembered across types/, sim/, domain/, platform/,
features/, and shell/.
