+++ playground/src/shell/index.ts
@@ -1,0 +1,3 @@
+export { boot } from "./boot";
+export type { State } from "./state";
+export type { UiHandles } from "./wire-ui";
